QUA-705 simplify pass: extract dispatchIngest, flatten exit-code
From /simplify findings:

- Extract evaluator → ingester dispatch into shared module
  crux/third-party-evals/ingesters/dispatch.ts. Both ingestSubcommand
  and backfillEvaluator now call dispatchIngest(). Adding a 5th
  evaluator now touches one place.
- Add EvaluatorKey string-union type and VALID_EVALUATORS constant
  for runtime validation messaging.
- Hoist duplicated `${r.error}: ${r.message}` into a single errorMsg
  variable.
- Replace nested-ternary backfill exit-code with named helper
  backfillExitCode() — flatter, easier to extend.
- Trim WHAT-comment header on backfill.ts to WHY-only (the four-step
  pipeline is already obvious from the imports).

Net: -94 lines / +26 lines across the two modified files; +43 LOC for
the new shared dispatch.ts module.

Skipped findings (deferred or out of scope):
- pLimit() duplication with citation-auditor.ts — only 2 instances,
  under the >3 threshold; touching citation-auditor.ts is QUA-692
  scope.
- ai-model-resolver.ts entityNormalizedForms() not memoized — this is
  in QUA-692's foundation, not the QUA-705 backfill driver. Would
  benefit a 200-report run but belongs in a separate PR/ticket.
